+++ CHANGES.md
@@ -1,3 +1,22 @@
+## Version 0.2.6:
+* Support of archlinux and debian install.
+* Support for any repos in debian.
+
+## Version 0.2.5:
+* Returning support of debian install. WIP.
+* Rewrting support of archlinux install.
+* Working on offline installation.
+
+## Version 0.2.4:
+* Split to two parts: profile generator and installer.
+* Added debug feature to own framework.
+* Fixed possible error with qemu emulation.
+* Minor fixes and cleanup.
+
+## Version 0.2.3:
+* Return support of debian installation.
+* Begin of making offline setup support.
+
 ## Version 0.2.2:
 * Fix fstab bug.
 * Temporaly removed support of debian installation.